I wanted to sync the sourceCode displayed in the webpage with a server file, but doing it naively leads to setSourceCode requests sent by the client to the server to be way too numerous (one per keystroke), and they compete which messed the server file. I tried to put a java swing timer, to only sens a setSourceCode request after 2.5 seconds of no keystrokes, but the client cannot use java library. I tried to use a javascript timer, but it doesn't look nice (and doesn't work).
Actually, I think there is no need for such a sync, just let the users edit the code in the browser, then hit a "submit" button to submit the code to the server, which can then process it and send back whatever he has to.
So next step is to add scalajs-react to the stack, to make easier to create the webpage of the application and add this button.